(tests) move android permissions tools in tests directory
